COM-1884 - Add mixin for line chart
